backend/models/gan_inference.py
import torch
import torch.nn as nn
from torchvision import transforms
from PIL import Image
import numpy as np

class GANInference:
    """Placeholder for future GAN-based features (StyleGAN, etc.)"""
    
    def __init__(self, device):
        self.device = device
    
    def generate_variations(self, image_path, num_variations=4):
        """
        Generate variations of an image
        (Simplified version - can be expanded with actual GAN)
        """
        image = Image.open(image_path).convert('RGB')
        variations = []
        
        # For now, create variations using transformations
        # In future, replace with actual GAN-based generation
        transforms_list = [
            transforms.Compose([
                transforms.ColorJitter(brightness=0.2, contrast=0.2, saturation=0.2, hue=0.1),
                transforms.RandomAffine(degrees=5, translate=(0.05, 0.05))
            ]),
            transforms.Compose([
                transforms.ColorJitter(brightness=0.3, contrast=0.3),
                transforms.RandomHorizontalFlip(p=1.0)
            ]),
            transforms.Compose([
                transforms.ColorJitter(saturation=0.3, hue=0.15),
                transforms.RandomRotation(10)
            ]),
            transforms.Compose([
                transforms.ColorJitter(brightness=0.15, contrast=0.25, saturation=0.25),
            ])
        ]
        
        for i in range(min(num_variations, len(transforms_list))):
            transformed = transforms_list[i](image)
            variations.append(transformed)
        
        return variations
    
    def blend_styles(self, style_paths, weights=None):
        """
        Blend multiple style images
        
        Args:
            style_paths: List of paths to style images
            weights: List of weights for each style (should sum to 1)
        """
        if weights is None:
            weights = [1.0 / len(style_paths)] * len(style_paths)
        
        assert len(style_paths) == len(weights), "Weights must match number of styles"
        assert abs(sum(weights) - 1.0) < 0.01, "Weights must sum to 1"
        
        # Load and blend images
        blended = None
        
        for path, weight in zip(style_paths, weights):
            img = Image.open(path).convert('RGB')
            img_array = np.array(img).astype(np.float32)
            
            if blended is None:
                blended = img_array * weight
            else:
                # Resize if needed
                if blended.shape != img_array.shape:
                    img = img.resize((blended.shape[1], blended.shape[0]))
                    img_array = np.array(img).astype(np.float32)
                
                blended += img_array * weight
        
        blended = np.clip(blended, 0, 255).astype(np.uint8)
        return Image.fromarray(blended)