backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Machine Learning
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional but useful
scipy==1.11.3
matplotlib==3.8.0